refactor: trpc logic
