+++ apps/blog/src/app/api/trpc/[trpc]/route.ts
@@ -1,7 +1,7 @@
 import { fetchRequestHandler } from "@trpc/server/adapters/fetch";
 
 import { createTRPCContext } from "~/server/context";
-import { appRouter } from "~/server/root";
+import { appRouter } from "~/server/router";
 
 const handler = (req: Request) =>
   fetchRequestHandler({
