+++ apps/blog/src/trpc/client.ts
@@ -2,7 +2,7 @@ import type { QueryClient } from "@tanstack/react-query";
 import type { Optional } from "ts-roids";
 import { createTRPCReact } from "@trpc/react-query";
 
-import type { AppRouter } from "~/server/root";
+import type { AppRouter } from "~/server/router";
 import { makeQueryClient } from "./query-client";
 
 export const trpcClient = createTRPCReact<AppRouter>();
