+++ apps/blog/src/trpc/server.ts
@@ -3,13 +3,14 @@ import "server-only"; // <-- ensure this file cannot be imported from the client
 import { cache } from "react";
 import { createHydrationHelpers } from "@trpc/react-query/rsc";
 
-import { createTRPCContext } from "~/server/context";
-import { appRouter } from "~/server/root";
-import { createCallerFactory } from "~/server/trpc";
+import { appRouter } from "~/server/router";
+import { createCallerFactory } from "~/trpc/trpc";
+import { createTRPCContext } from "./contex";
 import { makeQueryClient } from "./query-client";
 
 // IMPORTANT: Create a stable getter for the query client that
 //            will return the same client during the same request.
+
 export const getQueryClient = cache(makeQueryClient);
 
 const caller = createCallerFactory(appRouter)(createTRPCContext);
