+++ apps/blog/src/trpc/trpc.ts
@@ -1,8 +1,8 @@
 import { initTRPC } from "@trpc/server";
 import { ZodError } from "zod";
 
-import type { TrpcContext } from "./context";
-import { transformer } from "../trpc/transformer";
+import type { TrpcContext } from "./contex";
+import { transformer } from "./transformer";
 
 const t = initTRPC.context<TrpcContext>().create({
   transformer,
